Add files via upload
Simple adapter plate that will mount to the VORON Tap carriage and allow tthe Oribiter v3 to securely tie into the carriage.  This will cause you to loose 10mm of X travel.  I also included clearance holes for the carriage mounting bolts but these are not needed if on a CNC or CF TAP carriage.   Part cooling to be designed but should be able to mount a small fan on the left side fairly easily.